added endpoint protection

+++ pages/api/deck/delete.ts
@@ -1,5 +1,8 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import  prisma  from "../../../lib/prismadb";
+import { authOptions } from "../auth/[...nextauth]";
+import { getServerSession } from "next-auth/next";
+
 
 interface ExtendedNextApiRequest extends NextApiRequest {
     body : {
@@ -11,7 +14,15 @@ interface ExtendedNextApiRequest extends NextApiRequest {
 //     message : boolean
 // }
 
+
+
 export default async function handler(req: NextApiRequest, res: NextApiResponse){
+
+    const session = await getServerSession(req, res, authOptions);
+
+    if(session?.user?.id !== req.body.userId) {
+        return res.status(401).json("Unauthorized");
+    }
     const { id } = req.body;
 
     const deletedDeck = await prisma.deck.delete({
